Create Elements - Fix References Implementations
In this commit, I modified both the vertex and curv2D APIs so that the
handling of the vertex reference numbers and the curv2D reference
numbers are consistent with one another and work for all element types
that use either vertex reference numbers or curv2D reference numbers.
It is a little more verbose now, but at least the solution works for all
required scenarios, now.