Create Elements - Comments
In this commit, I added the functionality to the API for adding
comments.  The functionality is primarily provided by an interface,
Commentable, that provided the functionality to set and get comments.
And one method, "canComment()" that is currently in Node.java, but I
think I'll change that since it doesn't always make sense.