Create Elements - Refactor to Use Cursors
In this commit, I refactored the current API to use a Cursor object to
point to the target line in the document for any action instead of the
user specifying an integer value representing the line number.  I found
that the user providing the line number allowed for too many error
scenarios so introduced the line based cursor concept, instead.