Create Elements - Refactor to Use Strings in Factory
In this commit, I refactored the current solution so that the Factory
API uses primarily Strings as parameters except for maybe one or two
methods, even for primitives like booleans.  This is because the primary
intended user for the Factory is a parser.  The parser just parses out
the tokens.  I want the parser to have to do as little work as possible
otherwise.