Create Elements - Refactoring
In this commit, I refactored a number of items.  First, I refactored how
the cursor so that the operations to change the location of the Cursor
are entirely within Cursor interface as opposed to the Document
interface.  Second, I moved the operations to create the Document out of
the generic Factory and into its own separate Document Factory.  Then, I
renamed the generic Factory to StatementFactory since now all int
contains are operations related to creating statements, sub-statements,
etc., all associated with statements.  Finally, I refactored the
DocumentViewStatementCreationTests class to remove some redundant
statements and encapsulate the creation of the factories since I am
always tweaking things like that (for some reason).