Create Elements - Set Up Cursor Tests And Refactor
In this commit, I set up to write the Cursor tests and I also refactored
a little more to separate the Document related classes from the
Statement related classes.  It is intended that the Statement related
classes will be reused by the Logic view sub system so it makes sense to
keep them separate so that they do not become too coupled with the
Document classes.