Implement Cursor - Cursor Unit Tests
In this commit, to improve code coverage, I implemented some additional
unit tests for scenarios NOT defined in the acceptance tests.  The
acceptance tests' coverage was pretty good, though.  The only methods in
Cursor not covered in the acceptance tests were "equals()" and
"hashCode()", both methods whose contracts are defined by Object.
Although, precise definition for these methods are dependent on the
class being defined.