Implement Geometry - Refactor Common Methods Into Subclass
In this commit, the primary work item was to refactor the existing
solution to remove the common methods from the implementations of the
vertex and geometric statements into common base classes representing
statements and sub-statements.  This also required regenerating many of
the equals and hashCode implementations for the new class hierarchy.  I
also uncommented all acceptance tests for test scenarios that are not
yet implemented and added an @Ignore annotation.  I also removed
"Commentable" from the list of interfaces implemented by all statements
since the Statement interface already lists Commentable so it is
redundant.