Code Formatting
In this commit, I found an updated version of the Google Java Code
Formatting and applied it.  So, most of the changes are related to
reformatting.  Also, I changed one comment in the Document acceptance
testing class to reflect the fact that error scenario acceptance tests
should be added while working on the success scenarios.