Refactor Statements - StatementFactory Is Type Safe
In this commit, I modified the StatementFactory so that it is also at
the same level of abstraction as the Statement interfaces themselves,
thus, using the same types as defined in the accessors for the data
members in the Statement interfaces.  I also made some other minor
corrections along the way.