Refactor Statements - Comments Are Document Properties
In this commit, I modified how the comments are handled so that comments
are a property of Documents as opposed to being properties of
statements.  The rationale is that, conceptually, comments are not
really a property of a statement since they will not be used when
processing the OBJ statements at all.  Comments are, however, a property
of a document and improve the legibility of a document containing OBJ
data for human consumption.

This change substantially simplifies the statement class hierarchy,
which was kind of bothering me, anyway.

Decisions still need to be made about exactly how comments will be added
to documents, but it is not a big deal and I will delay this decision
until after more important work is complete.