Implement Statements - Statement Creation
In this commit, I implemented the statement creation and document
append, as they are presented in the
DocumentViewStatementCreationAcceptanceTests module.  The existing tests
are passing, however there are indeed errors in the implementation that
will be addressed when the statement unit testing is completed.  There
are also a couple statement implementations that warrant refactoring.