Review Statements - VertexReferenceGroup Fixes
In this commit, I implemented a number of fixes for VertexReferenceGroup
and VertexReference and their implementations.  Those issues include:
 - Missing implementations for hashCode() and equals().
 - Remove isSet from VertexReferenceImp constructor.
 - Ensuring lists are copied in a consistent manner as other statements.
 - Some other small fixes.

NOTE 1: Right now, I've got hashCode() and equals() defined in a number
of published interfaces.  This is likely not necessary but I'll leave it
for now since it is useful interface information.

NOTE 2: The current implementation for handling copying the
List<VertexReferenceGroup> is likely unsafe.  One of the design criteria
for this implementation is ensuring that the domain data is immutable
value objects.  Since the VertexReferenceGroup elements are NOT copied
from the source list to the destination list, it is feasible that a
mutable class instance that implements the VertexreferenceGroup
interface was slipped in.