Review Statements - Design Error Or Defensive Programming
In this commit, I removed an item from the README.md file that described
the copying of every statement before insertion into the document as a
design error.  It was not, in fact a design error.  It is a defensive
programming technique that is discussed in "Effective Java: Third
Edition" by Joshua Bloch, as item 50.  The design requires that all
statements are immutable value objects.  By copying the statements data
according to their published API into known statement implementations,
it protects against a malicious client that implements a mutable
statement that implements the published API and attempting to insert
that statement into the document.  I knew I did it for a reason ... I
just forgot why.  Hopefully, by writing it down, I'll remember for next
time.