Statement Review - Rename Document to DocumentView
In this commit, I renamed the module, and all references to the module,
from Document to DocumentView.  I did this because the DocumentView
implementation does not really have all the attributes someone would
expect a full document to have, such as spaces and comments.  Instead,
it is a logical view that represents a simplified representation where
the statements are organized in a document like fashion.  I may
implement a full Document at a later time.