Refactor Statements - VertexReference
In this commit, I've refactored VertexReference so that it is more type
safe.  Although the interfaces for the different types of vertex
references are currently all the same, the relationships between
VertexReferences and the statements that use them are not consistent
where some statements only support ParamVertexReference, or others only
support GeoVertexReference, etc.