Test Statements - Statement Acceptance Tests
In this commit, I implemented most of the acceptance tests for the
BasisMatrix based on the outline I provided previously.  I implemented
all tests related to the equals(), hashCode(), and compareTo() method
contracts in the same test method since they are all really closely tied
together and based on well documented contracts.