Test Statements - Curve2D Equals, HashCode, and CompareTo
In this commit, I implemented the tests for equals, hashCode, and
compareTo for the Curve2D statement.  I also implemented a generic
comparator that compares lists of objects that implement the comparable
interface.  The solution was based on what I implemented for Call, but
more generic.  Then I refactored the CallImp solution to use the same
generic comparator.