Review Tests - Unimplemented Tests and Malicious Copying
In this commit, I removed the unimplemented, default "test()" methods
from all acceptance tests since it was getting difficult to tell if
tests were failing because they were unimplemented or because there was
actually a problem.  There are over 500 tests now, afterall.  This left
me with a number of tests relate checking to ensure the immutability of
the value objects.  I implemented one prototype test for BasisMatrix to
test for this plus the fixes related to copying the Matrix object that
BasisMatrix uses.  It was at this point that I determined that
BigDecimal, a class used extensively in this project, has an
implementation issue as it is NOT a "final" class which means that
malicious subclasses for BigDecimal could be created that are NOT
immutable!

Fixing this would be a big issue, so it will take some time.