Event Reporting - Cleaning Up and Rewriting Description
In this commit, I rewrote the section for the common status reporting to
be a little more consistent and structured.  Note, that this
documentation is not intended to be permanent mostly because I don't
want the added work of maintaining the documentation as the system
evolves.  It is just intended to describe how I envision the system
working before I start the implementation.