Localizer - Implement Localizer Acceptance Tests
In this commit, I implemented the acceptance tests for the Localizer
published interface and the related code implementation.  This
implementation either highlighted or introduced a number of issues in
the current subsystem implementation that I'll need to take some time to
refactor.  Most of the work is listed in point form in the README.md
file, but I believe a number of automated tests are failing, too.  I'll
probably start the corrections there.