Localizer - Refactor ResourceBundle Creation
In this commit, I refactored the ResourceBundle creation so that both
the Factory and LocallizerBundle used the same code.  The implementation
is not what I originally intended, but this solution makes more sense
since the Factory is responsible for creating all instances and, more
importantly, handling all initialization error scenarios consistently.