Localizer - Ensure Internal API Refers To Internal Instances
In this commit, I refactored the code so that the Internal API
interfaces and the class implementations consistently refer to the
internal interfaces of class implementations for all data members and
methods.  For example, the classes must refer to the
LocalizerFieldInternal interface instead of LocalizerField.  This
required me to cast in a bunch of places, which is less than ideal, but
I have an idea to fix that, too.  The change may be a bit large, so I'll
make it in its own commit.