Localizer - Ensure Proper Level of Abstraction Used
In this commit, I went through the implementation to ensure that the
proper internal API used used within the system by all implementations.
This work highlighted a problem where the instances passed into the
published API are not checked to ensure they are the trusted
implementation.  At a later time, I'll need to implement safe casting to
ensure that the value object contract is preserved.