Localizer - SortedSet and Unmodifiable Set
In this commit, I implemented tests and the functionality to ensure that
the LocalizerBundle is returned as a SortedSet since it should be
because the LocalizerBundles are checked in the provided order and
should not have duplicates.  Also, I added tests to ensure that all sets
returned from the published API are unmodifiable.