Localizer - Safe Casting
In this commit, I implemented a solution to ensure that any instance
passed into the API by the client code is cast safely to the expected
internal type and throws an exception otherwise.  I probably chose the
wrong exception, but will fix that later.  I also reduced the
functionality considerably compared to what I intended but the solution
had in mind would likely cause side effects that I didn't want to deal
with for the sake of simplicity.