Localizer - Define System Published API
In this commit, I defined the System published API.  This is the start
of a solution I intend to implement to allow the same instances of
Localizer objects, especially LocalizerBundle instances, to be reused
across multiple clients.  The LocalizerBundle instances are particularly
important since they wrap ResourceBundle instances that cache results. I
am not sure whether these ResourceBundle instances are smart enough to
share caches across instances, I would assume so, but what value is
there in assuming?  As for the other Localizer classes, the
LocalizerType and LocalizerField instances are translated into string
tag names for localized strings within ResourceBundles.  Thus, they have
an inherent uniqueness constraint that must be managed by the
implementation.