Localizer - Localizer Refactor for Name and UID
In this commit, I refactored the Localizer implementation to so that the
Localizer instances can be uniquely identified by a given name.  I used
the same conventions as all other uniquely identified published domain
objects in the localizer subsystem for consistency.  Although the change
seems pretty large, since it is more of the same as before, the changes
were well defined and thus less risky.

I also fixed the System interface since it should've defined the method
interface getLocalizerKeySet() and not getLocalizerSet().