Localizer - Start ResourceBundle Handling Refactor
In this commit, I implemented the factory acceptance tests for the new
ResourceBundleWrapper module.  The purpose of the ResourceBundleWrapper
is, as its name implies, to wrap the actual ResourceBundle module
provided by Java.  This is being done for two reasons.  First, to better
define the interface between the Localizer and the ResourceBundle.  And
second, to more easily allow implementing stubs since testing the
Localizer functionality with the actual ResourceBundle implementation is
somewhat painful.  I'd rather use stubs and acceptance tests for the
ResourceBundle wrapper that explicitly demonstrate how the Localizer sub
system expects the ResourceBundle to behave.  You can say that it
indicates how a ResourceBundle must behave in order to be compatible
with Localizer.  This may also simplify introducing different
implementations of ResourceBundle in the future.