Localizer - Integrate ResourceBundleWrapper Into Localizer
In this commit, I replaced the ResourceBundle references in Localizer
with ResourceBundleWrapper.  As intended, it was a simple drop in
replacement.  I did not modify any of the existing tests in Localizer so
that I could ensure the overall localizer subsystem worked as before,
thus testing to ensure the integration was successful as far as the
existing acceptance and unit tests indicate.  The ResourceBundleWrapper
is a fairly thin wrapper, by design, so the fact I was successfully able
to replace the ResourceBundle is not too surprising.

Now the problem is I have a number of duplicate test cases between the
localizer subsystem and the ResourceBundleWrapper.  For all testing
purposes that involved the ResourceBundle, I will replace the actual
ResourceBundleWrapper instance with a stub whose behaviour can be set
during test set up.