Localizer - Refactored LocalizerFactory To Accept Stubs
In this commit, I refactored the LocalizerFactory so that it will accept
a factory to create stub LocalizerBundleWrapper instances for testing.
I also exposed the necessary internal methods in a
TestableLocalizerFactory class so that the testing can configure the
behaviour of the stub ResourceBundleWrapper as required for each
individual test.  I know how I intend to implement it, but it is not
implemented yet.