Localizer - Started Using ResourceBundleWrapperStub Instances
In this commit, I started implementing the changes to replace using the
production ResourceBundleWrapper instances with the
ResourceBundleWrapperStub instances but I did not get very far before
encountering a problem.  The problem is that I need to configure the
ResourceBundleWrapper for the root locale independently from the
resourceBundleWrapper for the target locale when testing the
CompositeLocalizerBundle.  After some consideration, it would likely be
easiest to provide this additional flexibility by pushing the
distinction between creating target and root locale
ResourceBundleWrapper instances down into the wrapper subsystem.  This
is a design trade-off since, strictly speaking, the currently defined
interface for the ResoureBundleWrapper provides a sufficient abstraction
for the localizer subsystem.  However, following the "design for
testability" principle, the above described solution is the easiest one
that I could think of.