Localizer - Integrate Create Root Locale Into Localizer
In this commit, I integrate the new functionality in WrapperFactory to
create a ResourceBundleWrapper specifically for the root locale into the
LocalizerFactory.  It was a minor change but one test case was broken
since it was using the StubWrapperFactory and the method is not
implemented there, yet.  It concerns me that more test cases aren't
broken ...