Localizer - Migrated Localizer Unit Tests To Use Stubs
In this commit, I migrated the Localizer unit tests to use the
configurable ResourceBundleWrapper stubs.  I also deleted some test
cases since the scenarios are sufficiently well covered by the
ResourceBundleWrapper unit tests.  I think that is the last of the tests
that needed migration and I should be able to remove the Localizer
resource bundles that I was (awkwardly) using for testing.