Localizer - Rename LocalizerMethod to LocalizerInstance
In this commit, I renamed the LocalizerMethod instance to
LocalizerInstance and made the method name a data attribute of the
LocalizerType. This makes more sense since the group (package), type,
and method are all attributes of the type whereas an instance is a
specific, well, instance of that type.

Right now, all tests are passing and the executable source code has been
updated to ensure that all references are correct.  But, the test source
needs to be updated to reflect the changed names.