Localizer - Rename LocalizerMethod to LocalizerInstance In Tests
In this commit, I went through all the acceptance and unit tests to
change any references from LocalizerMethod to LocalizerInstance.  I also
changed any reference from the LocalizerType instanceName to the
LocalizerType methodName, so it was all pretty confusing.  I might have
gotten all references but I would not bet on it.