Localizer - Handle Creating Duplicate Localizer Objects
In this commit, I started the implementation to properly handle creating
duplicate Localizer Objects.  For example, if we create a
LocalizerInstance with the same instance name twice, what happens the
second time?  The answer is that the existing instance is returned
assuming that the client intends to use the same LocalizerInstance in
multiple locations.  Most of the Localizer objects are implemented and
working, though some refactoring is necessary, except LocalizerBundle.
The test related to the LocalizerBundle is failing because, well, I'm
seeing some strange behaviour that I cannot currently explain.