Localizer - Fix Issue With Creating Same LocalizerBundle Twice
In this commit, I fixed the issue that broke the scenario where the
client tries to create the same LocalizerBundle twice.  The expected
behaviour is that the original LocalizerBundle is not added to the
Localizer's set of LocalizerBundles and the existing LocalizerBundle was
returned.  That wasn't happening.  Turns out the issue was a flaw in the
CompositeLocalizerBundle.equals() method.  Fixed.