UID - Refactor to Decouple UID And Localizer For Testing
In this commit, I refactored how the UID instances are created to use
the same structure as the other subsystems, i.e., a Factory.  Similar to
how the Localizer registers a factory for the ResourceBundle wrappers,
it will do the same for the UIDs.  Note, right now the Localizer is not
compiling because I removed the previous UID creation interface.  Once I
refactor Localizer, everthing should compile again.