UID - Fixing Broken Testcases
In this commit, I implemented fixes for broken testcases.  Now I have
MORE broken testcases then what I started with!  I have an idea what the
problem is and will fix it, but this is already a big change so want to
commit in case I need to back out the fix I intend to implement.