Localizer - Fixing Structural Problem
In this commit, I implemented the intended fixes to try and resolve the
structural problem that was causing the loop between constructing
certain instances.  So, instead of a 100+ testcase failures, I am down
to 9.  So the structural problem is resolved but I don't like the
solution, so I want to refactor it.  As before, since this is a pretty
large change, I am committing what I have now before making any more
changes.