UID - Fix More Tests
In this commit, I fixed the tests that were failing in the Localizer
subsystem because the uniqueness constraint on the UIDs from the uid
subsystem were causing the tests to fail.  This is an actual integration
problem.  And the first step in fixing it is decoupling the testing of
the uid and localizer subsystems by implementing a stub factory for
UIDs.