Message - Move Old Message Implementation
In this commit, I renamed the packages that contain the existing message
subsystem to identify them as the old implementation.  This is so that I
can use it as a reference as I implement the new message subsystem
without it getting in the way.  I also structured the identified work in
my notes a bit better.