Event Core - Base Implementation
In this commit, I took a different approach to starting this module.
Instead of first concentrating on creating all the types and validating
the creation of all the types, I instead implemented the base
proof-of-concept for the solution that demonstrates the fundamental
functionality.  Next, I'll take a look at all the additional test
scenarios that need to be implemented, including validation.