Event Core - Preparation and Refactoring
In this commit, I wrote out a rough list of test scenarios that I want
to implement in the EventCoreAcceptanceTests module.  I also found I
didn't like that the "registerSubscriber()" was published on the Channel
module and NOT the factory.  Since I am creating a more definitive break
between the Channel initialization and the Channel use, I decided it
makes more sense to have the newly renamed "addSubscriber()" method
published on the EventFactory so that all initialization activities can
occur from the EventFactory interface and all the Channel usage is from
the Channel, Publisher, and Subscriber interfaces.