Event Core - Add Enable Channel Feature
In this commit, I added the initial implementation for the feature to
enable channels to differentiate between when the Channel instance is
being initialized where events, publishers, and subscribers are being
created and registered versus when events are being processed.  The
former operations can only occur while the Channel is in its initial
state of disabled while the latter operations can only occur while the
Channel is enabled.  Thus, enabling a Channel is the final step of the
Channel initialization.