Event Core - Refactor to Separate Validation from Creation
In this commit, I refactored the EventFactory implementation to separate
out the validation logic from the creation logic.  This way there is
clear responsibility.  Also, it MIGHT allow alternate validation logic
to be swapped in at a later time.  Right now, the validation logic is
pretty sparse, but that won't last for long.