Event Core - EventFactory Validate Expected Naming Convention
In this commit, I implemented the tests to ensure that the EventFactory
validates that names used for Channels and Events use a consistent,
expected naming convention.  I also found one method that was not
checking to ensure its Channel parameter was not null so I added a test
for it.