Event Core - Basis But Flawed Cached Instances Implementation
In this commit, I implemented the basis for the cached instances for all
the event objects.  I have encountered some incompatibilities with the
UID subsystem that are not trivial to fix.  As such, I have decided to
stop using the UID subsystem with the event core subsystem.  After this
commit, I'll start refactoring the code to remove dependencies on the
UID subsystem.