Event Core - Completed Removal of UID Dependency
In this commit, I completed the removal of the UID dependency from the
event core subsystem.  All acceptance tests are passing, however, I
don't like the implementation since I needed to sacrifice some type
safety in the process.  Should not be a big deal.  I should be able to
add back the necessary type safety.