Event Core - Refactor To Move Subscriber Instance Handling
In this commit, I refactored the code so that The Channel instances now
get their subscriber lists from the ChannelCache instead of maintaining
their own list.  Two surprising revelations (to me anyway).  First, when
creating the instances, I need to pass in the reference to the top most
EventFactoryInternalXXXImp in the Decorator design pattern list.  I have
an implementation for this though I don't like it too much.  It seems to
work, though.  And second, I was not expecting the
EventFactoryInternalCreatorImp to not have to do anything when adding
subscribers.  Seems completely logical now, but was not expecting it
until the code was written.  Still have to move over the handling of the
publishers and events to the cache but I intend to take the exact same
approach.