EventCore - Refactor To Ensure Valid Channel When Enabling
In this commit, I implemented a fairly minor refactoring where I moved
the logic to enable a channel to the event factory cacher so that I
don't assume that the channel passed in as a parameter is a valid
channel instance created by this event system before attempting to
enable it.  More validation logic may be required but not all acceptance
tests are implemented yet so that is not surprising.