EventCore - A Couple Small Refactors
In this commit, I implemented two small refactorings.  First, in the
event factory cacher, I implemented a private method to return the
ChannelCache given a Channel.  Mostly just aesthetic.  And second, in
the ChannelCache, I implemented a "containsEvent()" method given the
Channel, event family, and event name.  Thus, the event factory does not
need to do as much work and is more decoupled from the ChannelCache.
And the ChannelCache reuses the same instance of EventForComparison,
thus not needing to instantiate an Event instance simply to check if it
exists in the cache or not.