EventCore - Implement Acceptance Tests For Adding Multiple Instances
In this commit, I implemented a couple acceptance tests for handling
when the client tries to add multiple instances of subscribers and
events.  I got to the test where I try and subscribe a subscriber
already subscribed to a different channel and found I would need to make
a fairly large structural change.  Thus, I decided to commit what I
currently have before refactoring the code for this particular test.