EventCore - Write Outline For Acceptance Tests For Event Subjects
In this commit, I wrote out the outline for the acceptance tests for the
Event Subject functionality.  This is an extension to the basic Event
where a Subject that the Event pertains to is included.  Thus, the same
Event can be published multiple times, just as long as the subjects are
unique.