EventCore - Implemented Three More Acceptance Tests
In this commit, I implemented three more acceptance tests.  Before
continuing, I would like to do a little bit of refactoring in the test
cases for how they handle the subscriber.  I'd like to implement a stub
subscriber that counts publish and unpublish events so I can get rid of
all the anonymous classes that do exactly that.