EventCore - Complete Implementation Of Subject Acceptance Tests
In this commit, I completed the implementation of the Subject acceptance
tests.  No code changes were required.  The test case implementations
are pretty ugly, though, so some refactoring of the tests to remove some
of the redundancy is likely in order.  However, I want to commit the
code changes I have now before starting.