EventCore - Refactor Acceptance Tests
In this commit, I refactored the acceptance tests.  First, I expanded
the AssertEventCore to centralize validating the state of events and
subjects for various scenarios so that the testing is more consistent.
Then, I modified the acceptance tests to use these new testing methods.
Second, I renamed the CountingSubscriberStub to
AccumulatorSubscriberStub mostly because the implementation didn't
really count anything so the name made no sense.