EventCore - Outlined Next Set of Acceptance Tests
In this commit, I outlined the next set of acceptance tests.  I also
implemented a small refactor in the EventFactoryInternalCreatorImp so
that the NoSubject singleton is no longer static.  There is no reason
for it to be static since it is not intended to be accessible in a
static context.