EventCore - Implement Several Refactorings
In this commit, I implemented several refactorings.  I renamed the
current channel "enable" to "open" since the terminology is more
consistent with channels.  Second, I renamed most references to
"eventChannel", "eventFamily", "eventName", and "channelName" to
"channel", "family", "name", and "name" where appropriate.  I changed
the implementation of Subject to remove "isDefined()" as it wasn't being
used for anything and I could not think of any scenarios where it would
be used.  And finally, I refactored the Subscriber implementation so
that subscribers also have names and it is their names that are the
basis for the implementations of "equals()", "hashCode()", and
"compareTo()".  However, it may be even safer if Subscribers are created
with external names that I could control.