EventCore - Continue To Refactor Event And EventDescription
In this commit, I continued to refactor Event and EventDescription.  In
this commit, it was mostly to ensure that references to event and
eventDescription were correct. I also found one error in the
AssertNaturalOrder implementation and fixed it.  I also deleted
EventInternalForComparisonImp.java since it is no longer being used.