Split Project - Fix Directory Structure
In this commit, I completed fixing the directory structure to
remove the extra wfparser3 subdirectory.  I never intended to
add it and just ignored it.  I was concerned about preserving
the module commit histories but it seems none of them are
preserved from changing the Java package names, anyway, so it
really does not matter.