Split Project - Clean Up README.md, Gradle, And LICENSE
In this commit, I cleaned up the README.md a little to get rid to
references to functionality that is now in their own projects.  More
clean up is necessary, though.  I also cleaned up the Gradle
configuration to ensure the dependencies reflect what is actually used
by this project.  And finally, I added an Apache License versin 2.0
LICENSE file.